Controller methods reorganized
